Move test down to webfluxServerTest and remove nestedPath option
