+++ instrumentation/spring/spring-webflux/spring-webflux-5.0/javaagent/src/test/java/server/base/ServerTestRouteFactory.java
@@ -113,7 +113,6 @@ public abstract class ServerTestRouteFactory {
                     })));
   }
 
-  @SuppressWarnings("deprecation") // testing instrumentation of deprecated class
   protected Mono<ServerResponse> respond(
       ServerEndpoint endpoint, BodyBuilder bodyBuilder, String body, Runnable spanAction) {
     if (bodyBuilder == null) {
