+++ testing-common/src/main/java/io/opentelemetry/instrumentation/testing/junit/http/HttpServerTestOptions.java
@@ -51,7 +51,6 @@ public final class HttpServerTestOptions {
   boolean testNotFound = true;
   boolean testPathParam = false;
   boolean testCaptureHttpHeaders = true;
-  boolean testNestedPath = false;
   boolean testCaptureRequestParameters = false;
   boolean verifyServerSpanEndTime = true;
 
@@ -174,12 +173,6 @@ public final class HttpServerTestOptions {
   @CanIgnoreReturnValue
   public HttpServerTestOptions setTestCaptureHttpHeaders(boolean testCaptureHttpHeaders) {
     this.testCaptureHttpHeaders = testCaptureHttpHeaders;
-    return this;
-  }
-
-  @CanIgnoreReturnValue
-  public HttpServerTestOptions setTestNestedPath(boolean testNestedPath) {
-    this.testNestedPath = testNestedPath;
     return this;
   }
